chore(market-matcher): rm dead matchers
